App/requirements.txt
numpy==1.16.6
pandas==1.1.5
streamlit==1.3.1
Pillow==8.4.0
findspark==1.4.2
py4j==0.10.9.2